requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.25.1
outcome==1.3.0.post0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.4.28
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tiktoken==0.6.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
vaderSentiment==3.3.2
wsproto==1.2.0
